requirements.txt
python-frontmatter~=1.0.0